Add preprocessable files to tests.
The preprocessor used is m4.
Use it as:
    m4 -Dllvm -I tests/runnable tests/runnable/fact.moc > /tmp/fact.moc
Then use mocc on /tmp/fact.moc.

I have set the extension of our "header" files to "hoc".
There are 3 "hoc" files now:
    - puts.hoc that contains the implementation of put_int in moc
      language and conditionally includes the implementation of put_char
      and put_string in the machine language;
    - puts-llvm.hoc and puts-tam.hoc that contains the implementation of
      put_char and put_string in the two machine language.

It would be cleaner to have a way to put the implementation in other
"moc" file but it is not obvious how to do such things with tam.

The factorial test does not return its result anymore. It prints it on
stdout. The expected output is in fact.moc.output.