When I was young, it seemed that life was so wonderful,
A miracle, oh it was beautiful, magical.
And all the birds in the trees, well they'd be singing so happily,
Joyfully, playfully watching me.
But then they send me away to teach me how to be sensible,
Logical, responsible, practical.
And they showed me a world where I could be so dependable,
Clinical, intellectual, cynical.